Create different app command structure, shift some duties to discordSvc
