--- bot/providers/msg_component_handlers.go
@@ -1,2 +1,0 @@
-package providers
-
